fetch-scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["Andrew Li <andrewli@u.northwestern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
amqp = "^5.3.1"
pika = "^1.3.2"
httpx = "^0.28.1"
selectolax = "^0.3.29"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
lxml = "^5.4.0"
playwright = "^1.52.0"
dotenv = "^0.9.9"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.7"
ruff = "^0.11.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
